switch to mmrcms! macro
