+++ src/rust/mmrcms/sources/gekkouhentai/src/lib.rs
@@ -1,15 +1,9 @@
 #![no_std]
-use aidoku::{
-	error::Result,
-	prelude::*,
-	std::{net::Request, String, Vec},
-	Chapter, DeepLink, Filter, Manga, MangaContentRating, MangaPageResult, MangaViewer, Page,
-};
-use lazy_static::lazy_static;
-use mmrcms_template::template::MMRCMSSource;
+use mmrcms_template::{mmrcms, template::MMRCMSSource};
+use aidoku::{MangaContentRating, MangaViewer};
 
-lazy_static! {
-	static ref INSTANCE: MMRCMSSource = MMRCMSSource {
+mmrcms! {
+	MMRCMSSource {
 		base_url: "https://hentai.gekkouscans.com.br",
 		detail_categories: "Categorias",
 		category_parser: |_, categories| {
@@ -32,35 +26,5 @@ lazy_static! {
 			}
 		},
 		..Default::default()
-	};
-}
-
-#[get_manga_list]
-fn get_manga_list(filters: Vec<Filter>, page: i32) -> Result<MangaPageResult> {
-	INSTANCE.get_manga_list(filters, page)
-}
-
-#[get_manga_details]
-fn get_manga_details(id: String) -> Result<Manga> {
-	INSTANCE.get_manga_details(id)
-}
-
-#[get_chapter_list]
-fn get_chapter_list(id: String) -> Result<Vec<Chapter>> {
-	INSTANCE.get_chapter_list(id)
-}
-
-#[get_page_list]
-fn get_page_list(id: String) -> Result<Vec<Page>> {
-	INSTANCE.get_page_list(id)
-}
-
-#[modify_image_request]
-fn modify_image_request(request: Request) {
-	INSTANCE.modify_image_request(request)
-}
-
-#[handle_url]
-fn handle_url(url: String) -> Result<DeepLink> {
-	INSTANCE.handle_url(url)
+	}
 }
