+++ src/rust/mmrcms/template/Cargo.toml
@@ -6,3 +6,4 @@ publish = false
 
 [dependencies]
 aidoku = { git = "https://github.com/Aidoku/aidoku-rs/" }
+lazy_static = { version = "1.4.0", features = ["spin_no_std"] }
